code/requirements.txt
groq>=0.4.0
rank-bm25>=0.2.2
rich>=13.7.0
python-dotenv>=1.0.0
langdetect>=1.0.9
tiktoken>=0.6.0
